Added coupled_tracert.py
Coupled tracert accepts a destination address (such as google.com) and
runs an individual traceroute for both IPv4 and IPv6. The results of the
traces are stored in a newly created text file with the pattern
<month>_<day>_<year>__<hour>_<minute> according to the current system
values. Note, the newly created text file is created in the location
from which coupled_tracert is executed.